Resolve #331 - moving disconnected champion to the fountain
*Disconnected player walks back to the fountain
*When player reaches fountain he got 99999 armor and magic resist

+++ GameServerCore/IGame.cs
@@ -1,5 +1,6 @@
 ﻿using GameServerCore.Maps;
 using GameServerCore.Domain;
+using System.Collections.Generic;
 
 namespace GameServerCore
 {
@@ -13,6 +14,6 @@ namespace GameServerCore
 
         bool SetToExit { get; set; }
 
-        bool HandleDisconnect(int userId);
+        bool HandleDisconnect(ulong userId);
     }
 }
